% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dotplot.R
\name{enrich_dotplot}
\alias{enrich_dotplot}
\title{My dotplot}
\usage{
enrich_dotplot(
  result_object,
  showcategory = 20,
  title_name = "Custom Gene Set Enrichment Results"
)
}
\arguments{
\item{result_object}{compareClusterResult object}

\item{showcategory}{A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.}

\item{title_name}{figure title}
}
\value{
plot
}
\description{
dotplot for enrichment result
}
\examples{
\dontrun{
results <- enrichRice(gene_ID = CO2_data,
                     geneset = "Top200",
                     ont = "ALL",
                     padjustMethod = "BH",
                     pvalueCut = 0.05,
                     qvalueCut = 0.2,
                     minSize = 5,
                     maxSize = 1000)

result_shift <- transform_result_to_dotplot(result_data = results)

enrich_dotplot(result_object=result_shift,
           showcategory = 20,
           title_name = "Custom Gene Set Enrichment Results")
}
}
