% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichRice.R
\name{enrichRice}
\alias{enrichRice}
\title{Custom Gene Set Enrichment Analysis}
\usage{
enrichRice(
  gene_ID,
  geneset = "Top200",
  ont = "ALL",
  padjustMethod = "BH",
  pvalueCut = 0.05,
  qvalueCut = 0.2,
  minSize = 10,
  maxSize = 500
)
}
\arguments{
\item{gene_ID}{a vector of gene id}

\item{geneset}{one of "Top200", "Top500", "Top1000"}

\item{ont}{One of "MG", "TG", and "DG" subontologies, or "ALL" for all three.}

\item{padjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{pvalueCut}{adjusted pvalue cutoff on enrichment tests to report}

\item{qvalueCut}{qvalueCutoff qvalue cutoff on enrichment tests to report as significant.}

\item{minSize}{minimal size of genes annotated for testing}

\item{maxSize}{maximal size of genes annotated for testing}
}
\value{
A \code{enrichResult} instance
}
\description{
用于将提供的基因集与Top200,Top500或者Top1000的Gene set富集，并且返回富集结果.
}
\examples{
\dontrun{
results<-enrichRice(gene_ID = CO2_data,
          geneset = "Top200",
          ont = "ALL",
          padjustMethod = "BH",
          pvalueCut = 0.05,
          qvalueCut = 0.2,
          minSize = 5,
          maxSize = 1000)
}
}
